requirements.txt
flask
flask-socketio
paho-mqtt
smbus2
adafruit-circuitpython-ina219
